--- app/views/events/rounds/maps/_map.html.haml
@@ -1,7 +1,0 @@
-.events-rounds-map{'data-target': 'events--round-map.map events--designated-lane.map'}
-  .round-map-loading{'data-target': 'events--round-map.loading_placeholder'}
-    %i.fa.fa-3x.fa-circle-notch.fa-spin
-    %p
-      Loading...
-
-= render 'events/maps/legend'
