--- app/views/events/rounds/maps/_penalty.html.haml
@@ -1,12 +1,0 @@
-:ruby
-  penalized ||= false
-  penalty_size ||= 0
-
-  classes = [
-    ('label' if penalized),
-    (penalty_size < 50 ? 'label-warning' : 'label-danger')
-  ].compact
-
-%span.round-map-competitor-penalty{class: classes}
-  - if penalized
-    = "-#{penalty_size}%"
