--- app/views/events/rounds/maps/_reference_point.html.haml
@@ -1,29 +1,0 @@
-- editable ||= false
-
-%span
-  - assignment = @round_map.reference_point_assignment(competitor)
-  - reference_point_presentation = capture do
-    %span{'data-target': 'events--competitor-reference-point.reference_point', 'data-id': (assignment.reference_point_id if assignment)}
-      - if assignment
-        = assignment.reference_point.name
-      - else
-        %i.fa.fa-fw.fa-ellipsis-h
-      - if editable
-        %i.fa.fa-fw.fa-caret-down
-
-  - if editable
-    .btn-group
-      %button.btn.btn-default.btn-xs.dropdown-toggle{'data-toggle': 'dropdown'}
-        = reference_point_presentation
-      %ul.dropdown-menu.dropdown-menu-right
-        %li
-          %a{href: '#', 'data-action': 'click->events--competitor-reference-point#clear_reference_point'}
-            %i.fa.fa-fw.fa-ellipsis-h
-        - @round_map.reference_points.each do |reference_point|
-          %li
-            %a{href: '#',
-               'data-reference-point-id': reference_point.id,
-               'data-action': 'click->events--competitor-reference-point#assign_reference_point'}
-              = reference_point.name
-  - else
-    = reference_point_presentation
