--- app/views/events/rounds/maps/_round_map.html.haml
@@ -1,79 +1,0 @@
-- can_edit = policy(@round_map.event).update? && !@round_map.event.finished?
-
-.fullscreen-container
-  .fullscreen-header
-    = link_to @event, class: 'btn-back' do
-      %i.fa.fa-chevron-left
-      = @event.name
-    \//
-    = t('events.show.round_map_view')
-
-  .fullscreen-body{'data-controller': 'events--round-map',
-                   'data-url': url_for,
-                   'data-dl-start': @event.designated_lane_start,
-                   'data-action': 'maps_api:ready@window->events--round-map#onMapsReady maps_api:failed@window->events--round-map#onMapsFailedLoad'}
-
-    .fullscreen-main-area{'data-controller': 'events--designated-lane',
-                          'data-action': 'round-map:show-dl@window->events--designated-lane#show_dl'}
-      = render 'events/maps/designated_lane_form'
-
-      = render 'events/rounds/maps/map'
-
-    .fullscreen-sidebar.round-map-sidebar
-      .fullscreen-sidebar-container.round-map-competitors
-        - @round_map.competitors_by_groups.each_with_index do |group, group_index|
-          .round-map-group
-            .round-map-group-title
-              = "#{t('events.rounds.map.group')} #{group_index + 1}"
-              %span.round-map-group-commands
-                %a{href: '#', 'data-action': 'click->events--round-map#toggleGroup'}
-                  Select
-            .round-map-group-list
-              - group.each do |competitor_data|
-                .round-map-competitor{'data-controller': 'events--competitor-reference-point',
-                    'data-url': event_round_reference_point_assignments_path(@round_map.event, @round_map.round, competitor_id: competitor_data.competitor_id),
-                    'data-competitor-id': competitor_data.competitor_id}
-                  .round-map-competitor-row
-                    %label.round-map-competitor-title
-                      %input{type: 'checkbox',
-                             checked: competitor_data.present? && group_index == 0,
-                             'data-competitor-id': competitor_data.competitor_id,
-                             'data-action': 'change->events--round-map#toggleCompetitor',
-                             'data-target': 'events--round-map.competitor'}
-                      %i.fa.fa-circle{style: "color: #{competitor_data.present? ? competitor_data.color : 'transparent'}"}
-                      %span.round-map-competitor-name
-                        = competitor_data.name.titleize
-                      = render 'events/rounds/maps/penalty',
-                               penalized: competitor_data.penalized,
-                               penalty_size: competitor_data.penalty_size
-
-                    .round-map-competitor-reference-point
-                      - if @round_map.show_reference_points?
-                        = render 'events/rounds/maps/reference_point',
-                                 competitor: competitor_data.competitor,
-                                 editable: can_edit
-                  .round-map-competitor-row
-                    .round-map-competitor-actions
-                      - if @round_map.show_reference_points?
-                        %a.btn.btn-xs.btn-link{'data-action': 'click->events--competitor-reference-point#show_dl'}
-                          Show DL
-                        = can_edit ? '|' : ''
-
-                      - if can_edit
-                        = link_to t('event_tracks.show.penalties'),
-                                  event_round_map_penalty_path(@round_map.event, @round_map.round, competitor_data),
-                                  remote: true,
-                                  class: 'btn btn-xs btn-link'
-                    .round-map-competitor-direction
-                      - if competitor_data.direction
-                        .events-rounds-map-direction.events-rounds-map-second-row
-                          %i.far.fa-compass
-                          = "#{competitor_data.direction.round}°"
-
-                      - exit_altitude = competitor_data.exit_altitude.to_i
-                      - highlight_exit = exit_altitude > 3810
-                      %span
-                        - if highlight_exit
-                          %i.fa.fa-exclamation-triangle.text-warning{'data-toggle': 'tooltip', title: 'Exit altitude SEEMS to be higher than allowed'}
-                        Exit:
-                        = exit_altitude
