--- app/views/events/rounds/maps/penalties/_modal.html.haml
@@ -1,14 +1,0 @@
-.modal-dialog.modal-lg
-  .modal-content
-    .modal-header
-      %button.close{type: 'button', data: {dismiss: 'modal'}}
-        %span{aria: {hidden: 'true'}}
-          &times;
-        %span.sr-only
-          Close
-      %h4.modal-title
-        = event_result_presentation(@result)
-
-    .modal-body
-
-      = yield
