--- app/views/events/rounds/maps/penalties/show.js.erb
@@ -1,6 +1,0 @@
-$('#modal').html(
-  '<%= j render partial: 'events/results/penalties/penalties_form',
-               layout: 'events/rounds/maps/penalties/modal',
-               locals: { show_cancel: true, result: @result, url: event_round_map_penalty_path(@event, @round, @result) } %>'
-);
-$('#modal').modal('show');
