--- app/views/events/rounds/maps/penalties/update.js.erb
@@ -1,15 +1,0 @@
-(function () {
-  $('#modal').modal('hide')
-
-  const selector = [
-    '.round-map-competitor[data-competitor-id="<%= @result.competitor_id %>"]',
-    '.round-map-competitor-penalty'
-  ].join(' ')
-
-  const partial =
-    '<%= j render 'events/rounds/maps/penalty',
-                 penalized: @result.penalized,
-                 penalty_size: @result.penalty_size %>'
-
-  document.querySelector(selector).outerHTML = partial
-})()
