--- app/views/events/rounds/maps/show.json.jbuilder
@@ -1,25 +1,0 @@
-json.competitors do
-  json.array! @round_map.competitors do |competitor_data|
-    next if competitor_data.empty?
-
-    json.id competitor_data.id
-    json.competitor_id competitor_data.competitor_id
-    json.name competitor_data.name.titleize
-    json.path_coordinates competitor_data.path_coordinates
-    json.start_point competitor_data.start_point
-    json.end_point competitor_data.end_point
-    json.after_exit_point competitor_data.after_exit_point
-    json.color competitor_data.color
-  end
-end
-
-json.place do
-  json.latitude @round_map.event.place.latitude
-  json.longitude @round_map.event.place.latitude
-end
-
-json.reference_points do
-  json.array! @round_map.reference_points do |reference_point|
-    json.extract! reference_point, :id, :name, :latitude, :longitude
-  end
-end
