+++ spec/controllers/events/rounds/maps_controller_spec.rb
@@ -1,4 +1,4 @@
-describe Events::Rounds::GlobeController do
+describe Events::Rounds::MapsController do
   it '#show' do
     event = events(:published_public)
     round = event_rounds(:distance_round_1)
